wip: refactoring mvc lib structure
